src/helpers/utils.ts
export const handleQuit = (stop: () => void) => {
  // Handle unexpected exits
  process.on('exit', stop);
  // Catches ctrl+c event
  process.on('SIGINT', stop);
  // Catches "kill pid" (for example: nodemon restart)
  process.on('SIGUSR1', stop);
  process.on('SIGUSR2', stop);
  // Catches uncaught exceptions
  process.on('uncaughtException', stop);
};

export const stopIntervals = (interval: NodeJS.Timeout, timeout: NodeJS.Timeout): void => {
  clearInterval(interval);
  clearTimeout(timeout);
};

export function pick<T extends object, K extends keyof T> (base: T, keys: K[]): Pick<T, K> {
  const entries = keys
    .filter(key => !!base[key])
    .map(key => ([key, base[key]]));

  return Object.fromEntries(entries);
}
